PDS-446 add `confirmButtonLoading` prop to `ConfirmationModal`
This commit adds a loading prop to to the `ConfirmationModal`'s confirm
button, to activate the button's loading state. This will be useful for
interactions where clicking the confirm button fires off an async action,
such as a a database delete operation, and we want to keep the user
in the modal until it completes.