(CDPE-3578) Add api for show/hide to <Input>
Currently, the trailing and leading icons on the Input component are only icons. This commit makes changes to the Input component api to allow a callback to be passed to the icon to facilitate a password show/hide for example. This includes a boolean to turn the icon into a button and then a prop so that the consumer can pass an onClick to the button.