(PDS-504) Add source control icons to react-components
This commit adds the four source control icons that we use in CD4PE as svgs to icons.js as per instructions at https://puppet.style/#/React%20Components/Icon. For each svg:
1. Design: they were already designed. So...
2. Engineering: ran each of them through svgo for minification
3. Engineering: did not need to strip out <svg> tags and svgo left only the path tags.
4. Engineering: removed fill attributes
5. Engineering: added each icon to icons.js. all except for azure devops (large) had a medium size.

One thing here that might be terrible was the decision to change the
'large' default viewbox to '0 0 32 32' because of the azure icon.
anything smaller and it was getting chopped off and if it's going to
work out of the box, this seemed like the lowest friction change. in the
browser, it didn't seem to actually change any of the other icons but
i'm not sure if something else didn't change that i'm not aware of or if
other icons might be changed in other scenarios. anyway, may need to
iterate on this part.