(PDS-502) Add scrolling to Tabs at smaller widths
Changes the default behavior of the Tabs component so that when the Tabs
container is too narrow to display all of its tabs horizontally, it
enables horizontal scrolling through the tabs and displays scroll
buttons at the right and/or left as appropriate.

This scrolling behavior can still be disabled by setting `scroll=false`
on the Tabs container.