(maint) Fix ability to npm link react-components
Prior to this commit, @puppet/react-components made use of the npm step
`prebuild` to delete the build directory prior to rebuilding the app.
If another development environment was consuming a linked version of
@puppet/react-components, this prebuild step caused the consuming
application to rebuild immediately, and crash because the build dir was
gone.

With this commit, we remove the prebuild step entirely. As of Webpack 4,
deleting the build dir is no longer necessary. Webpack deletes the
the dir itself when adding the new files. This allows all consuming apps
to utilize the linked version of @puppet/react-components during
develoment.

**Note:**
This is the only package within the design system that used the prebuild
step. I would expect linking to work as-is for all other packages,
though I did not test them.