(PDS-565) Allow values that are objects in OptionMenuList
In Arrowhead, we'd much like the ability to pass objects as values in our Select menus for the Add a Project page. This is because when you're selecting things in this particular form, you're not just selecting the name of something or a string, you're selecting an object that has many properties, all of which are required to get the next list of things. So instead of needing to take the selected string and go back to a list of objects to find the one we want, let's just allow the selection of objects directly in the component.

This change just adds PropTypes.shape({ key: ... }) with a key prop included to the 'value' prop, because when OptionMenuList renders the list, it uses the value as the React key but in the case of an object, we still need to have a key to give it.