(PDS-575) Adding the ability to choose the cancel button type similar to the current submit type prop
This PR covers the addition of the proptype cancelType will allow the user of the component to choose the button type for the cancel button. This PR also changes the label type of the tag component to type node as it can be used to render an icon or html element instead of just a string.