(PDS-511) Bugfix for autocomplete select not staying open when typing in uncontrolled
Previously In the open prop if you supply the boolean value true what we would expect to see is the autocomplete dropdown open the entire time - we are not seeing that it seems that is being ignored. This PR resolves this issue and fixes the Autocomplete Select dropdown not staying open when it is uncontrolled (for example when being consumed by a form).