(PDS-584) Add overlay component
Created overlay component that allows for content to be placed on top of
the current view. The overlay is positioned relative to a target
element. The final position of the overlay can be adjusted by changing
the values for `align` and `position`. Additionally, an offset can be
provided to completely override the position of the overlay.