(CDPE-4655) Add tooltip to Sidebar.Footer
Prior to this commit, if the Sidebar Footer was used and the sign-out
button was present, there was no way to add a tooltip to explain the
usage of this button. This makes it difficult for someone using a screen
reader to find this button and use it.

With this commit, we add an optional parameter to the Sidebar Footer to
specify tooltip text. If specified, the tooltip will appear above the
sign-out button when hovered over.