Update minimum Node version to 18, Webpack to v5 and Lerna to v6.
Update the rest of the package(s)  dependencies to the latest major version where possible. Some packages cannot be updated to latest due to other peer dependency requirements we have not yet upgraded such as React and React-Dom.

The updated stylelintand eslint versions required some changes in the components to match new style rules.

Remove as many redundant or unused dependencies as possible.